intial sign in with email workflow working
