+++ pages/api/get-email-by-id.tsx
@@ -2,37 +2,12 @@ import { DocumentData, getFirebaseTable, MemberEmail } from "@/lib/api";
 import { verifyAdminToken, verifyAuthHeader } from "@/lib/auth";
 import { FirebaseTablesEnum } from "@/lib/enums";
 import { db } from "@/lib/firebase";
+import { getEmails } from "@/lib/firebase-helpers/private/directory";
 import { doc, getDoc } from "firebase/firestore";
 import { NextApiRequest, NextApiResponse } from "next";
 
 async function getEmailById(userId: string): Promise<MemberEmail> {
-  const secureMemberData: DocumentData[] = await getFirebaseTable(
-    FirebaseTablesEnum.SECURE_MEMBER_DATA,
-  );
-  const emails = await Promise.all(
-    secureMemberData
-      .filter((secM) => secM.fields.email !== "")
-      .map(async (secM) => {
-        const docRef = doc(db, FirebaseTablesEnum.MEMBERS, secM.id);
-        try {
-          const docSnapshot = await getDoc(docRef);
-          if (docSnapshot.exists()) {
-            return {
-              id: secM.id,
-              email: secM.fields.email,
-              name: docSnapshot.data().name || null,
-              emailAbbr: docSnapshot.data().masked_email || null,
-              status: docSnapshot.data().status || null,
-              unsubscribed: docSnapshot.data().unsubscribed || false,
-            };
-          }
-        } catch (error) {
-          console.error(error);
-        }
-        return null;
-      })
-      .filter((email) => email !== null),
-  );
+  const emails = await getEmails();
   const email = emails.find((e) => e && e.id === userId);
   return email;
 }
